requirements.txt
Django==4.2.6
djangorestframework==3.14.0
djangorestframework-jwt
joblib
# django-mssql-backend
# mysqlclient
pymysql
# MySQLdb
drf-yasg
Pillow
requests
